ejercicios/entregas/eduardoLostal/EduardoLostal-EX001PooStudent/src/Student.java
public class Student {

    private String name;
    private String surname;
    private String Id;
    private float grade;

    /*
     * Creates a new student with the given name, surname, id and grade.
     * @param name The name of the student.
     * @param surname The surname of the student.
     * @param id The id of the student.
     * @param grade The grade of the student.
     */

    public Student(String name, String surname, String id, float grade) {
        this.name = name;
        this.surname = surname;
        Id = id;
        this.grade = grade;
    }

    public Student(){


    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getSurname() {
        return surname;
    }

    public void setSurname(String surname) {
        this.surname = surname;
    }

    public String getId() {
        return Id;
    }

    public void setId(String id) {
        Id = id;
    }

    public float getGrade() {
        return grade;
    }

    public void setGrade(float grade) {
        this.grade = grade;
    }
}
